% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{add_sequences}
\alias{add_sequences}
\title{Add sequences to svmu output}
\usage{
add_sequences(svmu_data, ref_fasta, query_fasta, mc.cores = 1)
}
\arguments{
\item{ref_fasta}{A character. The path to the file in fasta format
containing the sequence of the reference used for the alignment.
It will be automatically indexed by \code{\link[Rsamtools]{indexFa}}
if its .fai index does not already exist.}

\item{query_fasta}{A character. Same as ref_fasta but for the query sequence
used in the alignment.}

\item{mc.cores}{A numeric. The number of cores passed to \code{\link[parallel]{mclapply}}
for processing in parallel. The default is 1 (no parallelization).}
}
\value{
A data.frame with the same number of records (rows) as the input
svmu_data, but with added POS, REF and ALT columns. These columns correspond
to the columns with the same name in a VCF file and will eventually be used
as such in the final VCF file.
}
\description{
This function takes a data.frame of svmu SVs as input and returns
the same data.frame with the columns POS, REF and ALT appended to
it, thus preparing it for formatting to VCF
}
\examples{
NULL
}
