% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{extract_sequences}
\alias{extract_sequences}
\title{Extract the REF and ALT sequences of an SV}
\usage{
extract_sequences(svmu_data, ref_fasta, query_fasta)
}
\arguments{
\item{svmu_data}{A single row of a data.frame containing variants
identified by svmu. This function processes a single row because it
is meant to be called by \code{\link[base]{lapply}} or in parallel
by \code{\link[parallel]{mclapply}} on a completed data.frame. Typically
this function will not be called directly but instead will be called
by the wrapper \code{\link{add_sequences}}.}

\item{ref_fasta}{A character. The path to the file containing the reference
sequence in fasta format. The existence of an index is not checked
by this function because this is typically done by the wrapper
\code{\link{add_sequences}}.}

\item{query_fasta}{A character. Same as ref_fasta but for the query
sequence used in the alignment.}
}
\description{
This function takes a single SV (one row of a data.frame describing
svmu SVs) and returns the same SV with POS, REF and ALT columns
added to it. This function is meant to be called in parallel by
\code{\link{add_sequences}}.
}
\examples{
NULL
}
