% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversions.R
\name{flag_inversions}
\alias{flag_inversions}
\title{Flag SVs located in inverted regions}
\usage{
flag_inversions(svmu_data, inversions, maxgap = 0)
}
\arguments{
\item{svmu_data}{a data.frame returned by \code{\link{read_svmu}} or another
function returning a similar data.frame, such as \code{\link{process_insdup}}}

\item{inversions}{a GRanges object of inversions found in the dataset, as returned
by \code{\link{inverted_regions}}.}

\item{maxgap}{a numeric of length one. The maximum width of the gap between an SV
and an inversion for the SV to be considered as overlapping the SV region.
the value is 0 be default (immediately adjacent ranges are considered overlapping)
but this value should probably be a small positive integer in most cases to account
for the fact that e.g. some deletions may not overlap any of the ranges defined
as deletions.}
}
\value{
a data.frame similar to the input svmu_data but with an additional column
INVERTED which is a logical vector indicating whether the SV is located within
an inversion. A column called Q_OVERLAP is also added which indicates whether the
query alignments on either side of the SV overlap (TRUE) or NOT (FALSE).
}
\description{
This function takes a GRanges object of inverted regions between
two alignments and a data.frame of SVs identified by svmu and adds
a column indicating whether this SV is located within an inverted region.
Columns are also added indicating whether the alignments on either
side of the SV overlap or not.
}
\examples{
NULL
}
