% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{format_vcf}
\alias{format_vcf}
\title{Format svmu data for VCF output}
\usage{
format_vcf(
  svmu_data,
  sample_name,
  min_size = NULL,
  max_size = NULL,
  logging = TRUE
)
}
\arguments{
\item{svmu_data}{A data.frame with explicit sequences and position
as generated by the function \code{\link{add_sequences}}, with
POS, REf, and ALT columns.}

\item{sample_name}{A character. The name of the sample, for naming the
sample in the header line.}

\item{min_size.}{A numeric of length one. The minimum size (in absolute
value) for an SV to be kept. If NULL (the default), no filters are
applied.}

\item{max_size.}{A numeric of length one. The maximum size (in absolute
value) for an SV to be kept. If NULL (the default), no filters are
applied.}

\item{logging.}{A logical. Whether the user should be informed on the
number of markers removed at the SV size filtering step.}
}
\value{
A data.frame that constitutes the body of the VCF to output,
with the header line as column names and one record per row of the
data.frame.
}
\description{
This function takes a data.frame returned by \code{\link{add_sequences}}
as input and returns a data.frame compliant with VCF 4.3 format. This function
generates only the data records part of the VCF; the meta-informationheader
is generated by the function \code{\link{vcf_header}}.
}
\examples{
NULL
}
