% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversions.R
\name{inverted_regions}
\alias{inverted_regions}
\title{Get GRanges of inversions from svmu output}
\usage{
inverted_regions(filename, min_distance)
}
\arguments{
\item{filename}{a character, the name of the file of SVs output
by svmu. Typically "sv.txt"}

\item{a}{single numeric value. The minimum distance above which
two inversion ranges are not merged together.}
}
\value{
A GRanges object describing where inversions occur relative
to the sequence used as a reference in the alignment
}
\description{
This function takes a file of SVs output by svmu (typically "sv.txt")
and returns a GRanges object of inversions found in the data. Inversion
are reduced such that the returned object broadly describes regions that
are inverted between the two aligned genomes
}
\examples{
NULL
}
