% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insdup.R
\name{merge_insdup}
\alias{merge_insdup}
\title{Merge an insertion and its associated CNV record(s)}
\usage{
merge_insdup(svmu_data, min_size)
}
\arguments{
\item{svmu_data}{A data.frame containing an INS record and at least one
CNV record associated with it}

\item{min_size}{a numeric. Insertions smaller than this size are flagged
with INSDUP_SIZE as their SV_TYPE value.}
}
\value{
A single INSDUP record that represents the merged INS/CNV
record. Records that could be successfully merged are returned
with INSDUP as their SVTYPE value. Records that could not be successfully
merged are returned with either INSDUP_FAIL or INSDUP_SIZE depending
on the reason why it was not processed.
}
\description{
This function processes a single INSDUP record. It is meant to be called
from \code{\link{process_insdup}} using lapply  on the output of \code{\link{split_insdup}}
to process all such records simultaneously.
}
\examples{
NULL
}
