% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummerplot.R
\name{mummerplot}
\alias{mummerplot}
\title{Generate a mummerplot of the alignments at a variant's position}
\usage{
mummerplot(svmu_data, delta_file, expansion = 1, prefix = "", mp_path)
}
\arguments{
\item{svmu_data}{a single row from a data.frame returned by
\code{\link{read_svmu}} or \code{\link{process_insdup}}.
The function accepts a single row because it is meant to
be called on a full data.frame using \code{\link{lapply}}.}

\item{delta_file}{a character. The path to the delta file
describing the alignment as produced by MUMmer.}

\item{expansion}{The expansion factor for the plotting window around
the variant. By default (expansion = 1), the plotting window is
expanded by the length of the variant in both the x and y directions
along the reference and query sequences from the start and end positions
of the variant. This parameter can be adjusted to enlarge or shrink
the plotting window.}

\item{prefix}{character. A prefix that will be added to the output
file name in addition to the metadata on the variant. By default
the prefix is the empty string.}

\item{mp_path}{a character. The path to the mummerplot executable.}
}
\value{
\code{NULL}, invisibly. This function is called for its
side effect of writing a plot to disk.
}
\description{
This function takes a single row from a data.frame of svmu
SVs as returned by \code{\link{read_svmu}} and plots the
alignment between the reference and query genomes at that
position using the mummerplot command of the MUMmer suite
of tools. The plot is output to a PNG file and its name
contains information about the variant.
}
\examples{
NULL
}
