% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_svmu}
\alias{read_svmu}
\title{Reading the SVs output by svmu}
\usage{
read_svmu(filename, remove_duplicates = TRUE)
}
\arguments{
\item{filename}{a character, the name of the file (should be "sv.txt")}

\item{remove_duplicates}{logical. Whether duplicate entries in the
SV dataset should be removed. svmu appears to return at least some
CNVs in a duplicated fashion by outputting a record both for the
first part of the query sequence that overlaps it and the second part
of the query sequence. This option allows to keep only records that
have a unique combination of REF_CHROM, REF_START, REF_END, SV_TYPE,
ID and LEN. It therefore considers almost all fields but those related
to the query sequence as its coordinates may be different.}
}
\value{
a data.frame describing the SVs found by svmu
}
\description{
This function simply takes the sv.txt file output by svmu and
returns a data.frame of that data
}
\examples{
NULL
}
