% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insdup.R
\name{split_insdup}
\alias{split_insdup}
\title{split_insdup}
\usage{
split_insdup(svmu_data)
}
\arguments{
\item{svmu_data}{a data.frame of SVs output by svmu, as returned
by the function \code{\link{read_svmu}}.}
}
\value{
A list of data.frames bundling together the insertion and its
associated CNV(s). The list element with the name "0" contains records that
are not insertion/duplication events.
}
\description{
This function accepts a data.frame of SVs output by svmu, as returned
by the function \code{\link{read_svmu}}. Its purpose is to identify
the insertion records that are associated with CNVs (insertion duplications)
and output a list of data.frames bundling together the insertion and its
associated CNV(s). The list element with the name "0" contains records that
are not insertion/duplication events. These should not be processed further.
This function is typically called at the beginning of the \code{\link{process_insdup}}
function.
}
\examples{
NULL
}
