% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{svmu_to_vcf}
\alias{svmu_to_vcf}
\title{Convert and filter SVs output by svmu to VCF format}
\usage{
svmu_to_vcf(
  svmu_file,
  sample_name,
  ref_fasta,
  query_fasta,
  output_file,
  mc.cores = 1,
  logging = TRUE,
  remove_duplicates = TRUE,
  insdup = TRUE,
  min_insdup = 50,
  process_inversions = TRUE,
  min_inv_distance = 5000,
  inv_maxgap = 10,
  apply_filters = FALSE,
  min_size = NULL,
  sv_types = NULL,
  maxcov = NULL,
  max_size = NULL
)
}
\arguments{
\item{svmu_file}{A character. The path to the file containing the output
of svmu (typically "sv.txt")}

\item{sample_name}{A character. The name of the sample. Will be used as the
genotype column in the output VCF file and as a prefix for the VCF ID column.}

\item{ref_fasta}{A character. The path to the reference fasta file. Will be
indexed by \code{\link[Rsamtools]{faiIndex}} if the corresponding .fai index
does not already exist.}

\item{query_fasta}{A character. Same as \code{ref_fasta} but for the query sequence.}

\item{output_file}{A character. The VCF file to which output is directed.}

\item{mc.cores}{An integer. The number of cores to use for processing INSDUP
records and extracting REF/ALT sequences in parallel. By default, mc.cores = 1
which means that processing is not done in parallel.}

\item{logging}{Logical. Whether the user should be informed at various processing
steps (\code{TRUE} by default).}

\item{remove_duplicates}{Logical. Whether duplicate records in the input file
should be removed (\code{TRUE} by default). See \code{\link{read_svmu}} for more
details.}

\item{insdup}{Logical. Whether insertions that co-occur with duplications should
be processed and flagged as INSDUP (\code{TRUE} by default and recommended).}

\item{min_insdup}{An integer of length one. The minimum size of the insertion
part of an INSDUP record for it to be kept (50 by default). This filter is
only applied if \code{insdup} is \code{TRUE}.}

\item{process_inversions}{Logical. Whether inversions should be used to
flag SVs as located in inverted regions (\code{TRUE} by default and recommended).}

\item{min_inv_distance}{An integer. The minimum distance (in bp) between two inverted regions
for them to be merged together as a single one (see \code{\link{inverted_regions}})
for more details.}

\item{inv_maxgap}{An integer. The maximum distance separating the location of an SV from an
inverted region for this SV to be considered as located in an inverted region.
See \code{\link{inverted_regions}} for more details.}

\item{apply_filters}{A logical. Whether to apply the filters implemented by
\code{\link{filter_svmu}}. These comprise the parameters min_size, sv_types,
maxcov and max_size.}

\item{min_size}{An integer. The minimum size of an SV for it to be kept for further
processing. See \code{\link{filter_svmu}} for more details.}

\item{sv_types}{A character vector. The SV types to be kept for further processing.
See \code{\link{filter_svmu}} for more details.}

\item{maxcov}{A numeric. The maximum coverage on the reference and query for the SV
to be kept for further processing. See \code{\link{filter_svmu}} for more details.}

\item{max_size}{An integer. The maximum size for an SV to be kept for further processing.
See \code{\link{filter_svmu}} for more details.}
}
\value{

}
\description{
This function takes a file of SVs output by svmu (typically
called "sv.txt") and converts it to VCF format. This function
performs processing of combined insertion/duplications and
inversion flagging by default. No filters are applied by
default (i.e. \code{\link{filter_svmu}}) is not applied)
but is nevertheless strongly recommended.
}
\examples{
NULL
}
